Create pc_extension_test_rom.vhd
Created test program to demonstrate the use of the extension to read the pc. First the program performs a relative jump (jumps to pc+12). Then it returns, having stored pc+1 to the data register before the jump. This can be used to return from subroutines more easily without relying on the instructions staying in the same place in program memory.